refactor(web): route auth imports through modules/auth barrel
