+++ packages/web/src/app/[locale]/(auth)/register/page.tsx
@@ -1,4 +1,4 @@
-import { RegisterForm } from '@/modules/auth/RegisterForm'
+import { RegisterForm } from '@/modules/auth'
 
 export default function RegisterPage() {
   return <RegisterForm />
